requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
